Rename Panels construct to allow for more implementations
