+++ panel-android-sharedpreferences/src/main/AndroidManifest.xml
@@ -1,8 +1,8 @@
-<manifest package="inkapplications.switchgear.android"
+<manifest package="inkapplications.switchgear.panel.android.sharedpreferences"
     xmlns:android="http://schemas.android.com/apk/res/android">
     <application>
         <activity
-            android:name="switchgear.android.ControlPanelActivity"
+            android:name="switchgear.panel.android.sharedpreferences.SharedPreferencesPanelActivity"
             android:theme="@style/Theme.AppCompat.Light.NoActionBar" />
     </application>
 </manifest>
