+++ panel-android-sharedpreferences/src/main/java/switchgear/panel/android/sharedpreferences/ItemFactory.kt
@@ -1,4 +1,4 @@
-package switchgear.android
+package switchgear.panel.android.sharedpreferences
 
 import switchgear.MissingConfigException
 import switchgear.Parameter
