+++ panel-android-sharedpreferences/src/main/java/switchgear/panel/android/sharedpreferences/SharedPreferencesPanelActivity.kt
@@ -1,4 +1,4 @@
-package switchgear.android
+package switchgear.panel.android.sharedpreferences
 
 import android.content.Context
 import android.content.Intent
@@ -8,19 +8,19 @@ import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.xwray.groupie.GroupAdapter
 import com.xwray.groupie.kotlinandroidextensions.ViewHolder
-import inkapplications.switchgear.android.R
+import inkapplications.switchgear.panel.android.sharedpreferences.R
 import switchgear.Parameter
 import switchgear.provider.sharedpreferences.SharedPreferencesConfigProvider
 
 private const val PARAMETERS = "extra.parameters"
 
 /**
- * A list of overridable configuration parameters.
+ * A list of overridable configuration parameters in SharedPreferences.
  *
  * This allows you to view and override all of the specified configuration
  * parameters using the SharedPreferences provider.
  */
-open class ControlPanelActivity: AppCompatActivity() {
+open class SharedPreferencesPanelActivity: AppCompatActivity() {
     private val parameters get() = intent.getSerializableExtra(PARAMETERS) as Array<Parameter<Any>>
 
     private lateinit var controlPanel: RecyclerView
@@ -52,7 +52,7 @@ open class ControlPanelActivity: AppCompatActivity() {
 }
 
 fun Context.startControlPanel(parameters: List<Parameter<Any>>) {
-    Intent(this, ControlPanelActivity::class.java)
+    Intent(this, SharedPreferencesPanelActivity::class.java)
         .apply {
             flags = Intent.FLAG_ACTIVITY_NEW_TASK
             putExtra(PARAMETERS, parameters.toTypedArray())
