+++ panel-android-sharedpreferences/src/main/res/drawable/ic_restore_gray_900_24dp.xml
