+++ panel-android-sharedpreferences/src/main/res/layout/control_panel.xml
