+++ panel-android-sharedpreferences/src/main/res/layout/item_switch.xml
