+++ panel-android-sharedpreferences/src/main/res/values/strings.xml
